+++ lib/TaskScheduler/IdleCounter.h
@@ -6,12 +6,12 @@
 class IdleCounter
 {
 private:
-    unsigned long idleCounter = 0;
+    uint32_t idleCounter = 0;
     bool taskWasRun = false;
     RunnableInterface* taskScheduler;
     HardwareClockInterface* hwClock;    
-    unsigned long lastMeasurementTimeMicros = 0;
-    unsigned long maxIdleCountsPerSecond = 1e6; //large number
+    uint32_t lastMeasurementTimeMicros = 0;
+    uint32_t maxIdleCountsPerSecond = 1e6; //large number
 
 public:
 
@@ -19,7 +19,7 @@ public:
         taskScheduler(_scheduler), hwClock(_hwClock)
     {}
     
-    unsigned long getIdleCounter()
+    uint32_t getIdleCounter()
     {
         return idleCounter;
     }
@@ -29,12 +29,12 @@ public:
         return idleCounter * 100 / maxIdleCountsPerSecond;
     }
     
-    unsigned long getIdleCountsPerSecondAndResetCounter()
+    uint32_t getIdleCountsPerSecondAndResetCounter()
     {
-        unsigned long thisTime = hwClock->getMicrosecondsSinceBoot();
-        unsigned long timeDelta = thisTime - lastMeasurementTimeMicros;
+        uint32_t thisTime = hwClock->getMicrosecondsSinceBoot();
+        uint32_t timeDelta = thisTime - lastMeasurementTimeMicros;
         lastMeasurementTimeMicros = thisTime;
-        unsigned long thisIdleCounter = idleCounter;
+        uint32_t thisIdleCounter = idleCounter;
         idleCounter = 0;
         return thisIdleCounter * 1e6 / timeDelta;
     }
@@ -62,7 +62,7 @@ public:
     {            
         getIdleCountsPerSecondAndResetCounter();
         for ( int16_t i = 0; i < 1000; i++ ) {            
-            unsigned long thisTime = hwClock->getMicrosecondsSinceBoot();
+            uint32_t thisTime = hwClock->getMicrosecondsSinceBoot();
             taskScheduler->run(thisTime);            
         }
         maxIdleCountsPerSecond = getIdleCountsPerSecondAndResetCounter();
