+++ test/testTaskScheduler/testRecurringTaskGroup.cpp
@@ -53,7 +53,7 @@ void testRunAndIncrementsNextTime()
     RecurringTaskGroup<3> group( 1000*1000L );
     group.nextRunTimeMicros = 1000*1000L;
     
-    unsigned long thisTime = 1500*1000L;
+    uint32_t thisTime = 1500*1000L;
 
     TEST_ASSERT_TRUE( group.canRun(thisTime) );
     group.run(thisTime);
